Add casts to hopefully quiet an MSVC warning.

+++ Scripting/Scripting.cpp
@@ -359,7 +359,7 @@ bool Scripting::Execute(const std::string& strCommand,
       int i0, i1;
       SysTools::FromString(i0, strParams[0]);
       SysTools::FromString(i1, strParams[1]);
-      const int int_sz = m_vHistory.size();
+      const int int_sz = static_cast<int>(m_vHistory.size());
       p0 = (i0 < 0) ? 0 
                     : ((i0 >= int_sz) ? m_vHistory.size()-1 
                                       : static_cast<size_t>(i0));
@@ -394,7 +394,7 @@ bool Scripting::Execute(const std::string& strCommand,
       int i0, i1;
       SysTools::FromString(i0, strParams[1]);
       SysTools::FromString(i1, strParams[2]);
-      const int int_sz = m_vHistory.size();
+      const int int_sz = static_cast<int>(m_vHistory.size());
       p0 = (i0 < 0) ? 0 
                     : ((i0 >= int_sz) ? m_vHistory.size()-1 
                                       : static_cast<size_t>(i0));
